Enable compilation for ESP32 (no WiFi)
This commit configures PlatformIO to build the project for ESP32 and
resolves a few compilation issues that pop up for that configuration.

Note that WiFi has to be excluded because WiFiS3 is not available, but
that's fine. When I move to ESP32 the comms will have to be completely
different anyway.

+++ andromeda/src/effects.cpp
@@ -189,7 +189,10 @@ class SaturationGlow : public AbstractEffect
       cycleTime(NUM_STRIPS),  // this SHOULD call the default constructor of RandParam, picking 7 random values. I think.
       color(NUM_STRIPS)
     {
-      saturationAmplitude = min(saturationCenter, (255 - saturationCenter));
+      saturationAmplitude = min(
+        static_cast<byte>(saturationCenter),
+        static_cast<byte>(255 - saturationCenter)
+      );
     }
 
     void precompute(milliseconds t) override
